Setup load profile based indexes
