+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "pandas>=2.2.0",
     "entsoe-py>=0.7.10",
     "isodate>=0.7.2",
+    "pyxlsb>=1.0.10",
 ]
 
 [dependency-groups]
