+++ src/energy_cost/data/be/electricity/indexes/BelpexRLP0N.py
@@ -1,0 +1,1 @@
+FIRST_YEAR = 2022
