+++ src/energy_cost/index/__init__.py
@@ -1,6 +1,7 @@
-from .cached_entsoe_day_ahead_index import CachedEntsoeDayAheadIndex
+from .cached_index import CachedIndex
 from .dataframe_index import CSVIndex, DataFrameIndex, YAMLIndex
 from .entsoe_day_ahead_index import EntsoeDayAheadIndex
 from .index import Index
+from .load_profile_index import LoadProfileIndex
 
-__all__ = ["CachedEntsoeDayAheadIndex", "EntsoeDayAheadIndex", "Index", "DataFrameIndex", "CSVIndex", "YAMLIndex"]
+__all__ = ["CachedIndex", "EntsoeDayAheadIndex", "Index", "DataFrameIndex", "CSVIndex", "YAMLIndex", "LoadProfileIndex"]
